Amendments to student instructions
Auth0 signup process has changed, they now ask for a User Type.

Corrected the paths of two files:
1. `client/AddFruit.jsx` -> `client/components/AddFruit.jsx`
2. `client/SelectedFruits.jsx` -> `client/components/SelectedFruit.jsx`

Corrected the name of the `@auth0/auth0-react` module in two places.

Finished incomplete sentence about importing `useAuth0` hook

Minor clarification on the use of the `useAuth0` hook.

Corrected the `sub` property to `auth0Id` to match code sample.

Added languages to code blocks. This is minor but it means that you get
language based hilighting in github and in vscode.